requirements.txt
Flask==2.0.3
openai==0.27.0
python-dotenv==0.19.2
Werkzeug==2.0.3